Split the cairo callbacks out
